requirements.txt
Django
djangorestframework
flake8    
flake8-html
drf-nested-routers